Re-export Gen3 repo for calibration collection changes.
Because collection changes moved files around, this required a lot of
manual file moves, not just rerunning the the script.  In addition, the
script exported more defects than seem to be necessary; these were
trimmed out, but it would someday be worth investigating why they were
exported to begin with in case this hides a larger problem.  Right now,
that question doesn't seem worrisome enough to hold up the merge.